refactor: create map-related components folder
